feat: export getScaleFromCropRect
